single byte or str?
