+++ crates/ryo3-bytes/src/lib.rs
@@ -13,9 +13,13 @@ use pyo3::intern;
 use pyo3::prelude::*;
 mod anybytes;
 pub mod bytes;
-mod bytes_ext;
 mod bytes_like;
 
+#[cfg(feature = "multiple-pymethods")]
+mod pyo3_bytes;
+
+mod python_bytes_methods;
+mod ryo3_bytes;
 pub use crate::bytes::PyBytes;
 pub use bytes_like::{extract_bytes_ref, extract_bytes_ref_str};
